Crate logseq

logseq

Simple library to read/write to Logseq note graphs in Rust, built with comrak.

Usage

Read the docs online, or see the tests for example code.

Crate Features

regex-lite: use the regex-lite crate as the regular expression backend (enabled by default)

regex: use the regex crate as the regular expression backend

serde: enable support for serde serialization & deserialization of certain types

Support

Currently, the latest version of Logseq OG (file-based) is supported, with some caveats:

“Timetracking” should be disabled

“Preferred workflow” should be TODO/DOING

Graph reindexing is currently TODO, so journal entry files created through the library aren’t yet visible in Logseq (tracking issue)

Support for the DB version of Logseq is also TODO (tracking issue)

Contributing

Contributions are hugely appreciated! The repository on Codeberg is ideal, but the GitHub mirror works too :>

Modules

block

Block-level code, including views over blocks, like block::view::Task and block::view::DueBlock.

document

Document-specific code.

error

Error types.

graph

Graph code, providing the entrypoint for working with a graph directory.

prelude

Convenience module for glob imports.

properties

Module containing a common properties::Properties trait for block and document properties.