logseq

Module graph

Graph code, providing the entrypoint for working with a graph directory.

Structs

Graph

Representation of a Logseq graph.

GraphBuilder

A helper object for constructing a Graph.

GraphEntry

An entry in a Logseq graph.

Namespace

A Logseq page namespace. Used to build page filenames and represent a virtual directory structure for pages.

Enums

EntryKind

A type of entry in a Logseq graph, either journal or page.