logseq::properties

Trait Properties

pub trait Properties: TryFrom<FxHashMap<String, String>> {
    // Required methods
    fn take_properties(input: &mut String) -> Result<Self, Self::Error>;
    fn maybe_take_properties(
        input: &mut String,
    ) -> Option<Result<Self, Self::Error>>;
}

A common trait for property access. You probably want an implementor, like BlockProperties or DocumentProperties.

Required Methods

fn take_properties(input: &mut String) -> Result<Self, Self::Error>

Extract the properties from an input, removing them from said input. This returns an empty object if there were no properties.

Errors

This function only fails if an id property is present that can’t be parsed to uuid::Uuid.

fn maybe_take_properties( input: &mut String, ) -> Option<Result<Self, Self::Error>>

Extract the properties from an input, removing them from said input. If there were no properties, this returns None.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.